16_variant_binary_search: Complete variant implementations of binary search
Implement the following variants:
Look for the first node matching the value.
Look for the last node matching the value.
Look for the first node which is larger than or equals to value.
Look for the last node which is smaller than or equals to value.
Look for the node matching a value in a circular array. A bug
occurs when the start, middle and end value are the same in the
first loop. It cannot determine whether the boundary is.

Signed-off-by: Hu Ziji <davidhuzj@gmail.com>