24_binary_tee: Add sorted binary tree node delete function
Implement how to delete a node from binary tree. Not quite elegant
but can work.

Implement the reload step to reset the index and layer value in
each tree node and reassign nodes in to array. However, it is very
cost to iterate the tree twice in reload. The first iterate resets
the index and layer and finally get the number of layer. The second
one allocates new array if the number of layer changes and loads
tree nodes into new array. Lots of optimization work is required.

Signed-off-by: Hu Ziji <davidhuzj@gmail.com>